Fix an unexpected failure when deploying nested vms.
The CE Lab deployer communicate with assets via GCP Runtime Configurator
variables. There is a rare race condition when working with nested vms,
where the deployer tries to read the variable value before the assets
have a chance to set them. When this case occurs, the deployer assumes
that a catastrophic failure has occurred and aborts.

Failure log: http://shortn/_4DQUhcxF7F

But the failure is actually inocuous. This fix makes the deployer ignore
this error.

Change-Id: I6e022e665549f397f6ccca898eb51ad1507f30b3
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/2080606
Reviewed-by: Xiang Ji <jxiang@google.com>