Update EnableService to use the ServiceUsage Api.
As a result of Go deprecating Dep in favor of Module, CELab can no longer lock the version of its GoogleAPI dependencies. In the latest GoogleAPI, the enable API services has moved from ServiceManagement to ServiceUsage.

This change updates CELab to call ServiceUsage to enable APIs when setting up a project.

Also fixed a unit test that is blocking cl upload.

Bug: 1199031
Change-Id: Iabb916dba3e64cb0281712d10903ed0fb3887882
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/2827241
Reviewed-by: Xiang Ji <jxiang@google.com>