Add validity checks to prevent panic in Reflection ops.
A side effect of upgrading from Go 1.11.2 to 1.16.3 is that most CELab
Go dependencies are upgraded as well. Some of the GCP protos have
changed, and these in turn are breaking the the part of CELab that uses
reflection to copy these proto objects.

This fix adds a few additional validity checks to prevent the Reflection
code from panicking when copying invalid structures.

Bug: 1199031
Change-Id: Ie2413290fde4931614c6786c671cfce79444ff50
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/2842603
Reviewed-by: Xiang Ji <jxiang@google.com>